requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
langchain-text-splitters==0.0.1
tiktoken==0.6.0
voyageai==0.2.2
pinecone-client==4.0.0
